refactor: Move queries to a new file called queries instead of utils
