+++ src/ports/lists/constants.ts
@@ -1,0 +1,1 @@
+export const GRANTED_TO_ALL = '*'
